src/requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
cycler==0.10.0
docopt==0.6.2
flatbuffers==2.0
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
line-bot-sdk==2.0.1
Markdown==3.3.6
matplotlib==3.4.3
mediapipe==0.8.8.1
multidict==5.2.0
numpy==1.21.3
oauthlib==3.1.1
opencv-python~=4.5.3.56
opt-einsum==3.3.0
pandas==1.3.4
Pillow==8.4.0
pipreqs==0.4.11
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.1
PySide6==6.2.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
shiboken6==6.2.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
yarg==0.1.9
yarl==1.7.2
zipp==3.6.0
plyer~=2.0.0
pypresence~=4.2.0